FEAT[Monsters]: Implements monsters and major refactoring in collisions
- Implements monsters in the game but not yet interacting with player besides collision
- Rewriten some collision related code, mostly, in a more Rust way to treat some problems regarding mutating a collection item through its update method
and at the same time passing a mutable reference to the collection as argument. This works in Java but no in Rust and than the colission was changed to
be checked immutalbly and before the update code. The update method itself had the collections removed as it's now not needed.

monsters